Experiment: change to floats

+++ src/simplex-openmp.cpp
@@ -11,7 +11,7 @@
 
 using namespace std;
 
-struct Compare { double val; int index; };    
+struct Compare { float val; int index; };    
 #pragma omp declare reduction(minimum : struct Compare : omp_out = omp_in.val < omp_out.val ? omp_in : omp_out)
 #pragma omp declare reduction(maximum : struct Compare : omp_out = omp_in.val > omp_out.val ? omp_in : omp_out)
 
@@ -19,19 +19,19 @@ class Simplex {
 
   private:
     int m, n;
-    std::vector<std::vector<double>> A;
+    std::vector<std::vector<float>> A;
     std::vector<int> basic;    // size m.  indices of basic vars
     std::vector<int> nonbasic; // size n.  indices of non-basic vars
 
   public:
-    std::vector<double> soln;
-    double z;    // return value of the objective function.
+    std::vector<float> soln;
+    float z;    // return value of the objective function.
     int lp_type; // for return.  1 if feasible, 0 if not feasible, -1 if
                  // unbounded
 
-    const double INF; // unbelivably, C++ doesn't support static doubles
+    const float INF; // unbelivably, C++ doesn't support static doubles
                       // initialized in a class
-    const double EPS;
+    const float EPS;
     const static int FEASIBLE = 1; // int vars are ok though
     const static int INFEASIBLE = 0;
     const static int UNBOUNDED = -1;
@@ -49,8 +49,8 @@ class Simplex {
         Cycling is possible.  Nothing is done to mitigate loss of
         precision when the number of iterations is large.
     */
-    Simplex(int m0, int n0, std::vector<std::vector<double>> &A0,
-            std::vector<double> &B, std::vector<double> &C)
+    Simplex(int m0, int n0, std::vector<std::vector<float>> &A0,
+            std::vector<float> &B, std::vector<float> &C)
         : m(m0), n(n0), A(m0 + 1), basic(m0), nonbasic(n0), soln(n), INF(1e100),
           EPS(1e-9)
 
@@ -92,7 +92,7 @@ class Simplex {
 
         while (true) {
             int r = 0, c = 0;
-            double p = 0.0;
+            float p = 0.0;
             
             struct Compare max;
             max.val = p;
@@ -153,13 +153,13 @@ class Simplex {
             
             #pragma omp parallel
             {
-                double p_local = p;
+                float p_local = p;
                 int r_local = r;
 
                 #pragma omp for nowait
                 for (int i = 0; i < m; i++) {
                     if (A[i][c] > EPS) {
-                        double val = A[i][n] / A[i][c];
+                        float val = A[i][n] / A[i][c];
                         if(val < p_local) {
                             p_local = val;
                             r_local = i;
@@ -228,7 +228,7 @@ class Simplex {
     bool Feasible() {
         int r = 0, c = 0;
         while (true) {
-            double p = INF;
+            float p = INF;
             
             struct Compare min;
             min.val = p;
@@ -270,7 +270,7 @@ class Simplex {
             #pragma omp parallel for reduction(minimum:min)
             for (int i = r + 1; i < m; i++) {
                 if (A[i][c] > EPS) {
-                    double val = A[i][n] / A[i][c];
+                    float val = A[i][n] / A[i][c];
                     if (val < min.val) {
                         min.val = val;
                         min.index = i;
@@ -288,11 +288,11 @@ class Simplex {
 int main(int argc, char *argv[]) {
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
-    std::vector<std::vector<double>> A;
+    std::vector<std::vector<float>> A;
     int numRules = 20000;
     int numVars = 20000;
     std::mt19937 randGen(1);
-    std::uniform_real_distribution<double>randReal(0, 100000.f);
+    std::uniform_real_distribution<float>randReal(0, 100000.f);
 
     auto randFloat = [&](){return randReal(randGen) ;};
 
@@ -307,14 +307,14 @@ int main(int argc, char *argv[]) {
         }
     }
 
-    std::vector<double> B;
+    std::vector<float> B;
     B.resize(numRules);
     for (int i = 0; i < numRules; i++) {
         // std::cin >> B[i];
         B[i] = randFloat();
     }
 
-    std::vector<double> C;
+    std::vector<float> C;
     C.resize(numVars);
     for (int i = 0; i < numVars; i++) {
         // std::cin >> C[i];
